ioctl number is unsigned long, don't use int type here since it will be erroneously sign-extended on 64-bit platforms.
git-svn-id: https://svn.aros.org/svn/aros/trunk@54467 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/hidds/nouveau/drm/libdrm/arosdrmmode.c
@@ -31,7 +31,7 @@
 #define U642VOID(x) ((void *)(unsigned long)(x))
 #define VOID2U64(x) ((uint64_t)(unsigned long)(x))
 
-static inline int DRM_IOCTL(int fd, int cmd, void *arg)
+static inline int DRM_IOCTL(int fd, unsigned long cmd, void *arg)
 {
 	int ret = drmIoctl(fd, cmd, arg);
 	return ret < 0 ? -errno : ret;
